requirement.txt
uvicorn
jinja2
python-jose
passlib
bcrypt
aiofiles
starlette
sqlalchemy
fastapi
bcrypt==4.0.1
cryptography
psycopg2-binary
PyMySQL
python-multipart
passlib
httpx
pytest-asyncio
